chore: separate schema validation logic to a function
